Add override of calculate_field_in_all_records() that does not need the
2006-02-22  Murray Cumming  <murrayc@murrayc.com>

        * glom/base_db.cc:
        * glom/base_db.h: Add override of
        calculate_field_in_all_records() that does not need the primary key
        spoon-fed to it. Get all the primary key values, not all the calculated
        field values.
        * glom/mode_design/fields/box_db_table_definition.cc:
        on_Properties_apply(): Warn about necessary recalculation and
        allow the user to cancel the change of field definition.
        change_definition(): Recalculate all values if necessary. This
        even works.

+++ glom/mode_data/notebook_data.cc
@@ -231,7 +231,7 @@ Notebook_Data::dataview Notebook_Data::get_current_view() const
   const int current_page = get_current_page();
 
   dataview result = DATA_VIEW_Details;
-  if(current_page == m_iPage_List)
+  if(current_page == (int)m_iPage_List)
     result = DATA_VIEW_List;
 
   return result;
